% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmc.R
\name{rjmc}
\alias{rjmc}
\title{Reversible Jump Markov Chain Monte Carlo}
\usage{
rjmc(
  y,
  model,
  lambda,
  theta,
  kkappa,
  alpha,
  beta,
  alpha.kappa,
  beta.kappa,
  sigma,
  mu,
  run
)
}
\arguments{
\item{y}{Counting data}

\item{model}{Type of model}

\item{lambda}{The mean of Poisson distribution and negative binomial distribution}

\item{theta}{Prior distribution parameter}

\item{kkappa}{Negative binomial distribution parameter}

\item{alpha}{Prior distribution parameter}

\item{beta}{Prior distribution parameter}

\item{alpha.kappa}{Prior distribution parameter}

\item{beta.kappa}{Prior distribution parameter}

\item{sigma}{Normal distribution parameters constructed during model transfer}

\item{mu}{Invariants during model transfer}

\item{run}{Length of chain}
}
\value{
Model selection result
}
\description{
When modeling count data, one question that is often of interest is whether the data is too dispersed for a Poisson distribution. If it is too dispersed, it may be more appropriate to use a negative binomial distribution to fit, so for a given data y, the reversible jump MCMC method is used for model selection. That's what this function rjmc does.
}
