\name{SA23204192-package}
\alias{SA23204192-package}
\alias{SA23204192}
\docType{package}
\title{This package provides several functions for Reversible Jump Markov Chain Monte Carlo}
\description{\packageDescription{SA23204192}}
\section{Package Content}{\packageIndices{SA23204192}}
\author{
Zhiheng Qi
Maintainer: Zhiheng Qi <qizhiheng0208@163.com>
}
\keyword{package}
